Update packets serialization system and add new
Add padding marker (required to handle correctly TS_SC_SKILL)

Change hasNullTerminator to the only case seen : string in packet include
or not the null terminator. In any case, the size of dynamic string is the
real size in the packet.

Simplify handling of definition/implementation differences.

Add new packets and transform all GameClient to the serialization format.