Allow detaching the client without closing server connection
Move filter handling from client to server session
 - If client close its connection but server session is not closed,
   we need to keep filter states
 - If the server close its connection, the client session is always closed
 - Move handling of GS list in server session too

Add EndpointProxy so sending packets to the client while disconnected
is transparent.

Add onClientDisconnected and onServerDisconnected to allow filters to
handle the forwarding of the disconnection to the other side.
Filters can choose to not close the server connection if the client
connection is closed using onClientDisconnected.

Actually, onServerDisconnected won't be used, when the server is disconnected
the client is always closed too.
This simplify the handling of filters. They are owned by the server session.

Add tests to test rzfilter behavior:
 - Test normal foward of packets in both direction and via an auth
   connection and GS connection.
 - Test closing client or server without closing the other side
 - Rename the ping test
 - Use different packets for the ping test so the packet format is always valid