appveyor: prevent github release to cause build loops
If the tag name is not the same as the repo tag, a new one will be created
by github when creating the release. This will cause a new build on this
new tag in appveyor.

To prevent this, use the same tag for the release as the original built tag.